Spring boot backend web application
A blog is a website or section of a website where an individual or organization writes about topics of interest. Blogs are usually maintained by an individual with regular entries of commentary, descriptions of events, or other material such as images.

Its a backend web application to maintain these blogs. The flow starts with creating a user on the website. A user can write multiple blogs and update or delete its blogs. Each blog has some title, content and some (possibly multiple) images. The features of an image include its description, dimensions, etc.